Change detection rewritten
Reason: sometimes an infinite loop is created and it is difficult to
find the cause for it.

Caches are now immediately created, with promise attributes.
Earlier caches were created when promises were resolved.

Also writeGitStatus() now writes only blocks of status in an
animation frame. Sometimes writing a screen was slowing down
gradually.