Make scheduleUpdate() push schedule further
Scheduling worked so that if it was already pending, the existing
schedule was used. Now it works so that pending is canceled and
new schedule is created.

Potentially this pushes updates further. In a case where
scheduleUpdate() is called near the edge of start, new animation frame
might be the next one.

Also this separates scheduling from the actual update.