Fix illegal edit edge cases (reverting)
Reverting to last safe text was sometimes impossible.
If checkpoint is done within `insert` mode, it will vanish when
returning to `normal` mode. Also we must take care to which
version of checkpoint is used.

Also introduce `nextNormalMode()` which can be used to do
one time job, in `normal` mode (will be called immediately if we
already are in `normal` mode)